requirements.txt
jira
pymongo
requests
tiktoken
openai
anthropic
python-dotenv